Fix runtime errors on some compilers
+ The compiler was emiting wrong symbols during runtime which was caused due to strcat, from a stackoverflow
answer
"strcat needs to find the null terminator in the destination. Your *result points to uninitialised memory, which
happens to have a null terminator characters in. Appending` \0` just before combining the strings
should fix it. Also, you are not allocating enough space for the null terminator in variables".

+ Also add some more debug logs (will be removed later)

Fixes compilation on macOS default Clang